src-node/src/services/requirements.txt
openai-whisper==20231117
torch
numpy
ffmpeg-python 